% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{combine_sd_summary_table}
\alias{combine_sd_summary_table}
\title{Combine sd summary based on number of sample types}
\usage{
combine_sd_summary_table(result_dir, sample_num)
}
\arguments{
\item{result_dir}{A directory for result files}

\item{sample_num}{The number of samples}
}
\description{
Combine sd summary based on number of sample types
}
