% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{DEG_analysis}
\alias{DEG_analysis}
\title{This function is for identify DEG from a given matrix}
\usage{
DEG_analysis(expr_mat, group, thr_filter = log2(0.5), thr_FC = 2, thr_p = 0.05)
}
\arguments{
\item{expr_mat}{should a matrix in stead of a data frame, with column for samples and rows for signature(gene/protein/metabolics...)
For transcriptome, a matrix after log transformation is recommended.}

\item{group}{should be a factor whose length is identical to the number of the columns in expr_mat,
describing the group information of each column in expr_mat}

\item{thr_filter}{genes with average no larger than thr_filter in both groups will not be considered as DEGs}
}
\value{
A statistics table for deg
}
\description{
This function is for identify DEG from a given matrix
}
