% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_group_output.R
\name{get_more_group}
\alias{get_more_group}
\title{Get lst gene which was used to count snr by reference dataset}
\usage{
get_more_group(exp_fpkm_log, dt_meta, result_dir)
}
\description{
Get lst gene which was used to count snr by reference dataset
}
