% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_group_output.R
\name{get_one_group}
\alias{get_one_group}
\title{Get one group}
\usage{
get_one_group(
  dt_exp_melt,
  dt_exp_annot,
  exp_fpkm_log,
  dt_pairs,
  dt_meta,
  result_dir
)
}
\description{
Get one group
}
