% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_group_output.R
\name{make_performance_plot}
\alias{make_performance_plot}
\title{Make performance figure}
\usage{
make_performance_plot(
  dt_performance,
  xname = xname,
  yname = yname,
  gname = gname,
  file_name = file_name
)
}
\description{
Make performance figure
}
