% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_ref_data}
\alias{read_ref_data}
\title{Read reference data}
\usage{
read_ref_data(ref_data_dir)
}
\arguments{
\item{ref_data_dir}{A root directory for reference data}
}
\value{
The reference data as a list.
}
\description{
Read reference data
}
